% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeNoise.R
\name{DeNoise}
\alias{DeNoise}
\title{Denoise a reference Seurat Object}
\usage{
DeNoise(ref, lab, coef = 1.5)
}
\arguments{
\item{ref}{Reference Seurat object with pre-calculated UMAP}

\item{lab}{The reference label to treat as clusters. Default is "seurat_clusters"}

\item{coef}{The coefficient to multiply by the IQR. Default is 1.5}
}
\value{
Returns a Seurat object with a new slot called "denoise"
}
\description{
This function denoises, or removes sources of noise, which can affect downstream intergration analyses
}
\references{
Rheaume, B. A., & Trakhtenberg, E. F. (2022). Self-learning algorithm for denoising
and advancing the integration of scRNA-seq datasets improves the identification of resilient
and susceptible retinal ganglion cell types
\href{https://www.biorxiv.org/content/10.1101/2021.10.15.464552v2.abstract}{bioRxiv}.
}
