% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuclidAssign.R
\name{EuclidAssign}
\alias{EuclidAssign}
\title{Determine Euclidean Distance}
\usage{
EuclidAssign(query, ref, label)
}
\arguments{
\item{query}{Query UMAP}

\item{ref}{Reference UMAP}

\item{label}{The reference label to transfer to query cells}
}
\value{
Returns a data frame containing the distance to the nearest reference point for each query point and the label of that reference cell.
}
\description{
This function computes distances in multidimensional space and assigns variables based nearest neighbors
}
\examples{
# Create random test reference
test.ref <- data.frame(UMAP_1 = c(runif(5, min= -5, max = -3), runif(5, min= 5, max = 8), runif(5, min= 3, max = 6)),
                      UMAP_2 = c(runif(5, min = -5, max = -3), runif(5, min = 5, max = 6), runif(5, min= -1, max = 3))) 
# Create random labels for reference
Col <- as.factor(c(rep(1, 5), rep(2, 5), rep(3, 5))) 

# Create random test query
test.query <- data.frame(UMAP_1 = runif(5, min= -10, max = 10), UMAP_2 = runif(5, min= -10, max = 10)) 

# Plot cells with labels to view layout
pal <- rainbow(3)
plot(test.query, pch = 16, col = "black")
points(test.ref, pch = 16, col = pal[Col])

# Run EuclidAssign
EuclidAssign(query = test.query, ref = test.ref, label = Col)

}
\references{
Rheaume, B. A., & Trakhtenberg, E. F. (2021). Self-learning optimization algorithm for integration
of scRNA-seq datasets improves the identification of resilient and susceptible retinal ganglion cell types.
\href{https://www.biorxiv.org/content/10.1101/2021.10.15.464552v2.abstract}{bioRxiv}.
}
