% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalProbe.R
\name{GlobalProbe}
\alias{GlobalProbe}
\title{Optimized Tracing  of Query to Reference Data}
\usage{
GlobalProbe(dat, lab)
}
\arguments{
\item{dat}{A Seurat object}

\item{lab}{The reference @meta.data slot on which to compare global parameters}
}
\value{
Global correlation analyses of the clusters provided in lab parameter
}
\description{
This function examines the global transcriptomic properties of single cell data.
}
\references{
Rheaume, B. A., & Trakhtenberg, E. F. (2021). Self-learning optimization algorithm for integration
of scRNA-seq datasets improves the identification of resilient and susceptible retinal ganglion cell types.
\href{https://www.biorxiv.org/content/10.1101/2021.10.15.464552v2.abstract}{bioRxiv}.
}
