% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to.R
\name{map_to}
\alias{map_to}
\title{Map Query Data to Reference UMAP}
\usage{
map_to(ref, query, method = "id.score", range = c(5, 10))
}
\arguments{
\item{ref}{Reference Seurat object}

\item{query}{Query Seurat object}

\item{method}{The method to use for k-weight optimization. Options include: "id.score" (default), "sub-sample"}

\item{range}{The range of k-weight parameters to loop through; default: 5-10}
}
\value{
A Seurat object containing the transferred data labels stored in "predicted.ids," the
confidence scores saved in predicted.id.score, and the reference umap of the query object.
}
\description{
This function takes Seurat query and reference objects and identifies the optimal k-weight-parameter
and outputs the query object with updated metadata; see below.
}
\references{
Rheaume, B. A., & Trakhtenberg, E. F. (2021). Self-learning optimization algorithm for integration
of scRNA-seq datasets improves the identification of resilient and susceptible retinal ganglion cell types.
\href{https://www.biorxiv.org/content/10.1101/2021.10.15.464552v2.abstract}{bioRxiv}.
}
