% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTo.R
\name{MapTo}
\alias{MapTo}
\title{Map query Seurat object to reference Seurat object}
\usage{
MapTo(
  ref,
  query,
  lab = "seurat_clusters",
  k = "opti",
  get.anchors = T,
  dims = 30
)
}
\arguments{
\item{ref}{Reference Seurat object}

\item{query}{Wuery Seurat object}

\item{lab}{The reference label to treat as clusters. Default is "seurat_clusters"}

\item{k}{The k-weight to use when weighting anchors. Default is opti}

\item{get.anchors}{A logical that when true will first run Seurat's FindTransferAnchors. Default is TRUE.}

\item{dims}{Number of PCs to use when get.anchors = T. Default is 30}
}
\value{
Returns a Seurat object with a new slot where the k-weight is stored: misc$CellTools$opti_k
}
\description{
This function is a wrapper for Seurat's TransferData, IntegrateEmbeddings, and ProjectUMAP functions, meant to be run after
OptiK, as part of the Denoise/OptiK/MapTo pipeline.
}
\references{
Rheaume, B. A., & Trakhtenberg, E. F. (2022). Self-learning algorithm for denoising
and advancing the integration of scRNA-seq datasets improves the identification of resilient
and susceptible retinal ganglion cell types
\href{https://www.biorxiv.org/content/10.1101/2021.10.15.464552v2.abstract}{bioRxiv}.
}
