% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptiK.R
\name{OptiK}
\alias{OptiK}
\title{Find the optimal k-weight parameter for KNN}
\usage{
OptiK(
  data,
  lab = "seurat_clusters",
  range = c(10, 100),
  dims = 30,
  perc = 20,
  num = NA,
  seed = 1984
)
}
\arguments{
\item{data}{Reference Seurat object}

\item{lab}{The reference label to treat as clusters. Default is "seurat_clusters"}

\item{range}{The range of k-weight values to loop through. Default is 10-100.}

\item{dims}{The number of PCs to consider when discovering anchors}

\item{perc}{The percent of cells to subsample from the refeerence for the test query. Default is 20\%.}

\item{num}{If perc = F, OptiK will look for a specific number of cells specified here.}

\item{seed}{Set the seed for reproducible results. Default = 1984.}
}
\value{
Returns a Seurat object with a new slot where the k-weight is stored: misc$CellTools$opti_k
}
\description{
This function iterates through a select range of k-weight parameters and stores the optimal parameter in the Seurat object
}
