% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefOptiK.R
\name{RefOptiK}
\alias{RefOptiK}
\title{Optimization of K-Nearest Neighbors for Tracing of Query to Reference Data}
\usage{
RefOptiK(ref, query, lab, range = c(10, 100), ref.dims = 30, query.dims = 30)
}
\arguments{
\item{ref}{Reference Seurat object}

\item{query}{Query Seurat object}

\item{lab}{The reference @meta.data slot to transfer to query object}

\item{range}{The range of k-weight parameters to loop through; default: 10-100}

\item{ref.dims}{The number of PCs to use when creating the reference UMAP and discovering anchors.}

\item{query.dims}{The number of PCs to use when creating the query UMAP}
}
\value{
A Seurat object containing the transferred data labels stored in "predicted.ids," the
confidence scores saved in predicted.id.score, and the reference umap of the query object.
}
\description{
This function takes query and reference Seurat objects as input, identifies the optimal k-weight-parameter,
and outputs the query object with updated metadata and miscellaneous slots, based on arguments; see below.
}
\references{
Rheaume, B. A., & Trakhtenberg, E. F. (2021). Self-learning optimization algorithm for integration
of scRNA-seq datasets improves the identification of resilient and susceptible retinal ganglion cell types.
\href{https://www.biorxiv.org/content/10.1101/2021.10.15.464552v2.abstract}{bioRxiv}.
}
