% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to.R
\name{map_to}
\alias{map_to}
\title{Optimized Reciprocal Mappping of Query and Reference Data}
\usage{
map_to(
  ref,
  query,
  lab,
  method = "ref-sample",
  range = c(5, 100),
  sample = F,
  dims = 50
)
}
\arguments{
\item{ref}{Reference Seurat object}

\item{query}{Query Seurat object}

\item{lab}{The reference @meta.data slot to transfer to query object}

\item{method}{The method to use for k-weight optimization. Options include: "ref-sample" (default) or "id.score"}

\item{range}{The range of k-weight parameters to loop through; default: 5-100}

\item{sample}{If method = "id.score," the reference and query data will be downsampled to the sample size given by this parameter. If method = "ref-sample," the sample number will be the size of the "query" object split from the reference.}

\item{dims}{Number of PCs to use when creating the reference UMAP and when discovering anchors.}
}
\value{
A Seurat object containing the transferred data labels stored in "predicted.ids," the
confidence scores saved in predicted.id.score, and the reference umap of the query object.
}
\description{
This function takes query and reference Seurat objects as input, identifies the optimal k-weight-parameter,
and outputs the query object with updated metadata and miscellaneous slots, based on arguments; see below.
}
\references{
Rheaume, B. A., & Trakhtenberg, E. F. (2021). Self-learning optimization algorithm for integration
of scRNA-seq datasets improves the identification of resilient and susceptible retinal ganglion cell types.
\href{https://www.biorxiv.org/content/10.1101/2021.10.15.464552v2.abstract}{bioRxiv}.
}
